requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
numpy==1.23.5
pandas==1.5.3
pandasai==2.3.2
plotly>=5.17.0
seaborn>=0.12.0
matplotlib>=3.7.0
openai>=1.3.0
python-dotenv>=1.0.0
werkzeug>=2.3.0
openpyxl>=3.1.0
xlrd>=2.0.1
gunicorn>=21.0.0
PyYAML>=6.0